Move sigmoid to tools and add gt_unit_id to metrics df
